CoR: consolidate debug plan approval telemetry into localDebugPlanTelemetryUtils
