+++ src/webviews/copilotOnRails/extension/utils/localDebugPlanTelemetryUtils.ts
@@ -4,6 +4,8 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { maskUserInfo } from "@microsoft/vscode-azext-utils";
+import { CopilotOnRailsContext } from "../../../../utils/copilotOnRails/CopilotOnRailsContext";
+import { setCorProp } from "../../../../utils/copilotOnRails/telemetryUtils";
 import { findColumnIndex, findSection, findTable, flattenContent, isChecked, type LocalPlanData, type LocalPlanSection } from "../../views/utils/parseLocalDebugPlanMarkdown";
 
 /**
@@ -176,6 +178,27 @@ export function getLocalDebugPlanTelemetry(planData: LocalPlanData): LocalDebugP
         convenienceScriptOfferedCount: scripts.offered,
         convenienceScriptSelectedCount: scripts.selected,
     };
+}
+
+/**
+ * Records the PII-safe debug-plan summary properties that the "submit debug plan approval" action emits.
+ *
+ * This is the single implementation shared by both approval paths: the manual local-plan-view approval
+ * ({@link LocalPlanViewController}) and the autopilot auto-approval. Callers own opening the
+ * `submitDebugPlanApproval` telemetry action and setting `approvalOutcome`; this only derives and stamps
+ * the plan summary props. Any parsing error is swallowed and flagged via a `parseFailed` property so
+ * telemetry never blocks the approval flow.
+ */
+export function recordLocalDebugPlanApprovalTelemetry(context: CopilotOnRailsContext, planData: LocalPlanData): void {
+    try {
+        const telemetry = getLocalDebugPlanTelemetry(planData);
+        for (const [key, value] of Object.entries(telemetry)) {
+            setCorProp(context, `${LOCAL_DEBUG_PLAN_TELEMETRY_PREFIX}${key}`, value);
+        }
+    } catch {
+        // Telemetry extraction must never block the approval flow; swallow any parsing errors.
+        setCorProp(context, `${LOCAL_DEBUG_PLAN_TELEMETRY_PREFIX}parseFailed`, true);
+    }
 }
 
 //#region Section metrics
